src/backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
#backoff==2.2.1
#beautifulsoup4==4.14.2
clip @ git+https://github.com/openai/CLIP.git
pytz==2025.2
#coloredlogs==15.0.1
#contourpy==1.3.2
#cryptography==46.0.3
#cycler==0.12.1
#dataclasses-json==0.6.7
#Deprecated==1.3.1
#deprecation==2.1.0
#dirtyjson==1.0.8
#distro==1.9.0
#dnspython==2.8.0
#emoji==2.15.0
#et_xmlfile==2.0.0
#exceptiongroup==1.3.0
fastapi==0.116.1
#filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
#ftfy==6.3.1
#geomet==1.1.0
#gitdb==4.0.12
#GitPython==3.1.45
#greenlet==3.2.4
#grpc-google-iam-v1==0.14.2
#h11==0.16.0
#h2==4.3.0
#hpack==4.1.0
#httpcore==1.0.9
#httpx==0.28.1
#httpx-sse==0.4.3
#huggingface-hub==0.36.0
#humanfriendly==10.0
#hyperframe==6.1.0
#idna==3.11
#iopath==0.1.10
#Jinja2==3.1.6
#jiter==0.11.1
#joblib==1.5.2
#jsonpatch==1.33
#jsonpointer==3.0.0
#jsonschema==4.25.1
#jsonschema-specifications==2025.9.1
#kiwisolver==1.4.9
langchain==0.3.27
langchain-astradb==1.0.0
langchain-community==0.3.27
langchain-core==0.3.79
langchain-ollama==0.3.9
langchain-text-splitters==0.3.9
langchain-unstructured==0.1.6
langdetect==1.0.9
langsmith==0.3.45
layoutparser==0.3.4
#lxml==6.0.2
#MarkupSafe==3.0.3
#marshmallow==3.26.1
matplotlib==3.10.7
#ml_dtypes==0.5.3
#mpmath==1.3.0
multidict==6.7.0
#mypy_extensions==1.1.0
#narwhals==2.10.2
#nest-asyncio==1.6.0
#networkx==3.4.2
#nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.0
onnx==1.19.1
onnxruntime==1.19.2
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.4
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==9.10.2
pillow==11.3.0
pillow_heif==1.1.1
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==22.0.0
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pymongo==4.15.3
PyMuPDF==1.26.4
pyparsing==3.2.5
pypdf==4.3.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
#shapely==2.1.1
#six==1.17.0
#smmap==5.0.2
#sniffio==1.3.1
#soupsieve==2.8
#SQLAlchemy==2.0.44
starlette==0.47.3
#striprtf==0.0.26
#sympy==1.14.0
#tabulate==0.9.0
#tenacity==8.5.0
#threadpoolctl==3.6.0
#tiktoken==0.8.0
#timm==1.0.22
#tokenizers==0.22.1
#toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
#tornado==6.5.2
#tqdm==4.67.1
#transformers==4.57.1
#typing-inspect==0.9.0
#typing-inspection==0.4.2
#typing_extensions==4.15.0
#tzdata==2025.2
unstructured==0.14.5
unstructured-client==0.42.3
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
watchdog==6.0.0
#wcwidth==0.2.14
#wrapt==2.0.1
##xlsxwriter==3.2.9
yarl==1.22.0
#zstandard==0.23.0
Authlib==1.6.5
itsdangerous>=2.0
pyjwt==2.10.1
extra-streamlit-components == 0.1.81
rank_bm25
